+++ src/pages/user/sekai_profile/SekaiID.tsx
@@ -250,6 +250,7 @@ const SekaiID: React.FC<unknown> = observer(() => {
     autorun(() => {
       setLocalSekaiProfile(sekaiProfileMap.get(region));
     });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {