+++ src/pages/user/sekai_profile/SekaiUserCardList.tsx
@@ -153,6 +153,7 @@ const SekaiUserCardList = observer(() => {
     autorun(() => {
       setLocalSekaiCardTeam(sekaiCardTeamMap.get(region));
     });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {